# docs/swagger.yaml
definitions:
  meson_msg.SaveFileMsg:
    properties:
      nameHash:
        type: string
      originUrl:
        items:
          type: string
        type: array
      sizeLimit:
        type: integer
    type: object
host: spec-00-02-015-bchkakakbxxxxxx-019-thisisthebinddomain.mesontrackingdomain.com
info:
  contact:
    email: contact@meson.network
    name: Meson Support
    url: https://meson.network
  description: meson terminal's api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://meson.network
  title: Meson Terminal API
  version: "1.0"
paths:
  /anyPath:
    get:
      description: get cached file
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: error msg
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: get cached file
      tags:
      - public
  /api/delete/:nameHash:
    get:
      description: delete file on terminal disk
      parameters:
      - description: 0dea69026ee1c698
        in: path
        name: nameHash
        required: true
        type: string
      - description: sdfwefwfwfwfsdfwfwf
        in: header
        name: Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: '{"msg": "who    are  you"}'
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: delete file command
      tags:
      - server cmd
  /api/health:
    get:
      description: health check
      produces:
      - text/plain
      responses:
        "200":
          description: UTC time string
          schema:
            type: string
      summary: health check
      tags:
      - public
  /api/nodestatus:
    get:
      description: get node status
      parameters:
      - description: sdfwefwfwfwfsdfwfwf
        in: header
        name: Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: error msg
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: get node status
      tags:
      - server cmd
  /api/pause/:second:
    get:
      description: pause file transfer for several seconds, and do speed test
      parameters:
      - description: "4"
        in: path
        name: second
        required: true
        type: string
      - description: sdfwefwfwfwfsdfwfwf
        in: header
        name: Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: error msg
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: pause file transfer
      tags:
      - server cmd
  /api/restart:
    get:
      description: restart node command
      parameters:
      - description: 0dea69026ee1c698
        in: path
        name: nameHash
        required: true
        type: string
      - description: sdfwefwfwfwfsdfwfwf
        in: header
        name: Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: error msg
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: restart node command
      tags:
      - server cmd
  /api/save:
    post:
      consumes:
      - application/json
      description: save file from given url
      parameters:
      - description: save command object
        in: body
        name: SaveFileMsg
        required: true
        schema:
          $ref: '#/definitions/meson_msg.SaveFileMsg'
      - description: sdfwefwfwfwfsdfwfwf
        in: header
        name: Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: '{"msg": "who    are  you"}'
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: save file command
      tags:
      - server cmd
  /api/schedulejobstatus:
    get:
      description: check ScheduleJob running status
      parameters:
      - description: 0dea69026ee1c698
        in: path
        name: nameHash
        required: true
        type: string
      - description: sdfwefwfwfwfsdfwfwf
        in: header
        name: Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg": "hello  Razeen"}'
          schema:
            type: string
        "400":
          description: error msg
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: check ScheduleJob running status
      tags:
      - server cmd
  /api/test:
    get:
      description: test api
      produces:
      - text/plain
      responses:
        "200":
          description: UTC time string
          schema:
            type: string
      summary: test
      tags:
      - public
  /favicon:
    get:
      description: handle favicon request
      produces:
      - text/plain
      responses:
        "200":
          description: empty string
          schema:
            type: string
      summary: handle favicon request
      tags:
      - public
schemes:
- https
swagger: "2.0"